fix(Golpe): corrigir fatoDeDano de int para double.

+++ Modulo01/Golpe.java
@@ -1,9 +1,9 @@
-public class Golpe {
+public class Golpe{
 
     private String nome;
-    private int fatorDano;
+    private double fatorDano;
 
-    public Golpe(String nome, int fatorDano) {
+    public Golpe(String nome, double fatorDano) {
         this.nome = nome;
         this.fatorDano = fatorDano;
     }
@@ -12,7 +12,7 @@ public class Golpe {
         return this.nome;
     }
 
-    public int getFatorDano() {
+    public double getFatorDano() {
         return this.fatorDano;
     }
